Refactor feed comments and bets into files
